Refactor Meshes and OBJ loading
A mesh is now templated based on MeshType::Simple or MeshType::Full.
Similarly, the OBJ loading is also now templated on the MeshType.
For `Simple` meshes, only geometry (triangles) are loaded in and
put into a BVH.

For `Full` meshes (Not implemented yet), the idea is that all the
sub-meshes from the OBJ file are loaded in separately, and the
original TriangleMesh<Full>::mesh is just a BVH of the sub-meshes.
In effect this gives us a 3-tiered system for Full meshes:

  - Global Object BVH (contains all objects)
    - Full Mesh BVH (contains sub meshes)
      - Sub Mesh BVH (contains triangles)