Add better support for reading .mtl files
The loader now reads in more of the values for the material and
uses them (really badly for now) to guess the type of material.

The choices being made right now are (very) arbitrary and definitely
not right in general, but they seem to work OK for the scenes I'm
testing on.