Add RescaledMesh class
This is simply a TriangleMesh equivalent, except that the resulting
mesh is by default scaled to fit inside a unit cube centered at
the origin. In some cases where mesh sizes can vary greatly, this
is convenient to have a consistent size. Use it the same way as
a TriangleMesh (including the <MeshType> template)