Medium: Correctly set `rec.obj`
Like with CSGs, now you can manually set a material onto the actual
ConstantVolume object, and if it exists `rec.obj` will point to the
the ConstantVolume object.

Otherwise, we set `rec.obj` to be the object that was returned by
`m_obj->hit(...)`. Earlier, we were hardcoding `rec.obj = m_obj`,
but this does not necessarily make sense if `m_obj` doesn't have it's
own material (for instance a CSGObject)