Medium: Add initial prorotype of Generic Volumes
Some ray marching is being done here to perform volumetric
intersections. Unlike in ConstantMedium, here the Medium object
keeps track of a DensityTexture, which it can sample for density
information at a given point.

For debugging, a ConstantDensityTexture has been implemented, as
well as a PerlinDensityTexture. Note that the same `density` values
for ConstantMedium and ConstantDensityTexture do not mean the same
thing.