Scene: Add a global pointer to the current scene
This is a dirty hack to allow us to add lightsources from parsed
OBJ files (coming soon in a commit near you...). Currently there's
no mechanism to be able to modify the scene from inside objects.

In the future perhaps we should modify the finalize() function to
take a referenace to the current scene to be able to perform
modifications to it.