All: Replace unnecessary includes with forward declarations
Compile times were getting a bit annoying, so most unnecessary
includes that were easy to get rid of have been replaced by
forward declarations.

Some files need a lot of forward declarations, which is a bit of
an eyesore, so in the future maybe there should be a `core/forward.h`
with forward declarations of most of the common things we need,
so we can just include that instead.

Fresh compiles are now ~35% faster!