Triangle/Mesh: Flip the X axis when loading the mesh
For some reason most meshes loaded seem to be flipped on the X
axis, and previoulsy we just flipped the `u` texture coordinate
for the triangle to fix flipped textures. The geometry was still
wrong.

Now, we just flip the mesh itself along the X, and can remove
the `1-u` hack from the triangle's texture coords generation.