HitRec: Add separate RayTypes for reflection/refraction
Earlier we only distinguised between diffuse / specular, but it
is useful to have a third option to be able to also distinguish
when we want to refract, for the (upcoming) `Hybrid` material.

For now this is fine, but in the future if we add more material
types we may just want to add a `float seed` field in the `HitRec`
class, and then have the `BSDF` instances figure out how to
interpret the random seed value generated by `BSDF::initSample()`.