Materials: Add Hybrid Material
This is a convenient material with parameters that allow us to
perform different amounts of diffuse, reflection and refraction.
It can also be configured to have perturbed specular bounces.

While not physically accurate, it is nice to be able to show a wider
range of materials without having to dynamically switch between
BSDFs.