DirectLighting/Path: Move out `SampleLight` function
Instead of duplicating the explicit light sampling code in multiple
places, we just move the `SampleLight()` method outside the `Path`
class (there wasn't a need for it to be a class method anyway).

Now, the `DirectLighting` renderer can just use this function. We
now add the ambient light inside `DirectLighting::Li()`