LoadedMedium: Optimize density checking
Checking the density data array is extremely critical, so having
clamping operations for the index bounds is expensive. Instead, we
just scale up the ray position from the canonical [0-1] box to
`dim-2` insead of `dim-1`, this way the trilinear interpolation
never goes out of bounds. The effect on scale is negligible.`